+++ core/src/main/scala/geotrellis/store/GeoTrellisRasterSourceLegacyProvider.scala
@@ -1,0 +1,15 @@
+
+package geotrellis.store
+
+import geotrellis.raster.RasterSourceProvider
+import geotrellis.raster.geotiff.GeoTiffPath
+
+class GeoTrellisRasterSourceLegacyProvider extends RasterSourceProvider {
+  def canProcess(path: String): Boolean = {
+    (!path.startsWith(GeoTiffPath.PREFIX) && !path.startsWith("gdal+")) &&
+      path.nonEmpty &&
+      GeoTrellisPath.parseOption(path).nonEmpty
+  }
+
+  def rasterSource(path: String): GeoTrellisRasterSourceLegacy = new GeoTrellisRasterSourceLegacy(path)
+}